llm-terminal-test-bundle/requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
rich>=13.7.1
jsonschema>=4.22.0
